requirements.txt
# SQL Agent TUI - Requirements
# Python 3.11+

# TUI Framework
textual>=0.47.1
rich>=13.7.0

# AI/LLM Dependencies - Updated to latest versions
langchain>=0.2.0
langchain-google-genai>=1.0.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langgraph>=0.2.0
langchain-core>=0.2.0

# Database - PostgreSQL
psycopg2-binary>=2.9.9

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.3

# For API requests to fetch models dynamically
httpx>=0.27.0
google-generativeai>=0.3.0
